Tests: Components - simplify directory structure
