+++ Tests/CMakeLists.txt
@@ -49,36 +49,28 @@ if (OGRE_BUILD_TESTS)
     endif ()
 
     if (OGRE_BUILD_COMPONENT_PAGING)
-      include_directories(${CMAKE_CURRENT_SOURCE_DIR}/Components/Paging/include)
       set(OGRE_LIBRARIES ${OGRE_LIBRARIES} OgrePaging)
-      list(APPEND HEADER_FILES Components/Paging/include/PageCoreTests.h)
-      list(APPEND SOURCE_FILES Components/Paging/src/PageCoreTests.cpp)
+      list(APPEND SOURCE_FILES Components/PageCoreTests.cpp)
     endif ()
     if (OGRE_BUILD_COMPONENT_MESHLODGENERATOR)
-      include_directories(${CMAKE_CURRENT_SOURCE_DIR}/Components/MeshLodGenerator/include)
       set(OGRE_LIBRARIES ${OGRE_LIBRARIES} OgreMeshLodGenerator)
-      list(APPEND HEADER_FILES Components/MeshLodGenerator/include/MeshLodTests.h)
-      list(APPEND SOURCE_FILES Components/MeshLodGenerator/src/MeshLodTests.cpp)
+      list(APPEND SOURCE_FILES Components/MeshLodTests.cpp)
     endif ()
     if (OGRE_BUILD_COMPONENT_TERRAIN)
-      include_directories(${CMAKE_CURRENT_SOURCE_DIR}/Components/Terrain/include)
       set(OGRE_LIBRARIES ${OGRE_LIBRARIES} OgreTerrain)
-      list(APPEND HEADER_FILES Components/Terrain/include/TerrainTests.h)
-      list(APPEND SOURCE_FILES Components/Terrain/src/TerrainTests.cpp)
+      list(APPEND SOURCE_FILES Components/TerrainTests.cpp)
     endif ()
     if (OGRE_BUILD_COMPONENT_PROPERTY)
-      include_directories(${OGRE_SOURCE_DIR}/Components/Property/include)
       set(OGRE_LIBRARIES ${OGRE_LIBRARIES} OgreProperty)
-      list(APPEND SOURCE_FILES Components/Property/src/PropertyTests.cpp)
+      list(APPEND SOURCE_FILES Components/PropertyTests.cpp)
     endif ()
     if (OGRE_BUILD_COMPONENT_OVERLAY)
       set(OGRE_LIBRARIES ${OGRE_LIBRARIES} OgreOverlay)
     endif ()
     
     if(TEST_GLSUPPORT)
-      include_directories(${OGRE_SOURCE_DIR}/RenderSystems/GLSupport/include)
       set(OGRE_LIBRARIES ${OGRE_LIBRARIES} OgreGLSupport)
-      list(APPEND SOURCE_FILES RenderSystems/GLSupport/src/GLSLTests.cpp)
+      list(APPEND SOURCE_FILES RenderSystems/GLSupport/GLSLTests.cpp)
     endif()
     
     if(ANDROID)
